Move AdaptConfigCx, AdaptEventCx to new module; make ctors public
